Roleauth

+++ Controllers/FamilyController.cs
@@ -1,14 +1,14 @@
 using Microsoft.AspNetCore.Mvc;
 using YourVillage.Models;
 using System.Collections.Generic;
-using System;
 using System.Linq;
-using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using System.Threading.Tasks;
 using System.Security.Claims;
+using YourVillage.Authorization;
+using System;
 
 namespace YourVillage.Controllers
 {
@@ -18,25 +18,55 @@ namespace YourVillage.Controllers
   {
     private readonly YourVillageContext _db;
     private readonly UserManager<ApplicationUser> _userManager;
-    public FamilyController(YourVillageContext db, UserManager<ApplicationUser> userManager)
+    private readonly IAuthorizationService _authService;
+    public FamilyController(YourVillageContext db, UserManager<ApplicationUser> userManager, IAuthorizationService authService)
     {
       _db = db;
       _userManager = userManager;
+      _authService = authService;
+
     }
 
     public async Task<ActionResult> Index()
     {
       var userId = this.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
       var currentUser = await _userManager.FindByIdAsync(userId);
-      var userFamily = _db.Families.Where(entry => entry.ParentUser.Id == currentUser.Id);
-      ViewBag.Children = new List<Child>();
-      foreach (Family family in userFamily)
+      var userFamilies = (_db.Families.Where(entry => (entry.ParentId == currentUser.Id))).ToList();
+
+      if (_db.Caregivers.Select(entry => (entry.CaregiverId == currentUser.Id)) != null)
       {
-        var child = (_db.Children.Where(entry => entry.FamilyId == family.FamilyId));
-        ViewBag.Children.Add(child);
+
+        var famout = new List<Family>();
+        var careFamilies = _db.CaregiverFamilies
+        .Where(c => c.CaregiverId == currentUser.Id);
+        if (careFamilies.Count() != 0)
+        {
+          foreach (CaregiverFamily cf in careFamilies)
+          {
+            famout.Add(_db.Families.FirstOrDefault(f => f.FamilyId == cf.FamilyId));
+          }
+        }
+        ViewBag.CareFamilies = famout;
       }
-      return View(userFamily);
+
+      ViewBag.Children = new List<Child>();
+      foreach (Family family in userFamilies)
+      {
+        var child = (_db.Children.Where(entry => entry.FamilyId == family.FamilyId)).ToList();
+        ViewBag.Children = child;
+      }
+
+      foreach (Family family in userFamilies)
+      {
+        var babysitters = _db.Families
+        .Include(fam => fam.Caregivers)
+        .ThenInclude(join => join.Caregiver)
+        .FirstOrDefault(fam => fam.FamilyId == family.FamilyId);
+      }
+      return View(userFamilies);
     }
+
+
 
     public ActionResult Create()
     {
@@ -48,16 +78,79 @@ namespace YourVillage.Controllers
     {
       var userId = this.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
       var currentUser = await _userManager.FindByIdAsync(userId);
-      ViewBag.ParentUser = currentUser;
+      family.ParentId = currentUser.Id;
       _db.Families.Add(family);
       _db.SaveChanges();
       return RedirectToAction("Index");
     }
 
-    public ActionResult Details(int id)
+    public async Task<ActionResult> Details(int id)
     {
       Family thisFamily = _db.Families.FirstOrDefault(family => family.FamilyId == id);
+      var isAuthorized = await _authService.AuthorizeAsync(User, thisFamily, YourVillageOperations.Read);
+      if (!isAuthorized.Succeeded)
+      {
+        return Forbid();
+      }
+      var babysitters = _db.Families
+      .Include(fam => fam.Caregivers)
+      .ThenInclude(join => join.Caregiver)
+      .FirstOrDefault(fam => fam.FamilyId == id);
       return View(thisFamily);
+    }
+
+    public async Task<ActionResult> AddCaregiver(int id)
+    {
+      Family thisFamily = _db.Families.FirstOrDefault(family => family.FamilyId == id);
+      var isAuthorized = await _authService.AuthorizeAsync(User, thisFamily, YourVillageOperations.Update);
+      if (!isAuthorized.Succeeded)
+      {
+        return Forbid();
+      }
+      return View(thisFamily);
+    }
+
+    [AllowAnonymous]
+    public async Task<ActionResult> CaregiverAccess(bool isCaregiver)
+    {
+      var userId = this.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+      var currentUser = await _userManager.FindByIdAsync(userId);
+      if (_db.Caregivers.Where(c => (c.CaregiverId == currentUser.Id)).ToList().Count != 0)
+      {
+        return View();
+      }
+      else
+      {
+        return RedirectToAction("Create", "Caregiver");
+      }
+    }
+
+    [AllowAnonymous]
+    [HttpPost]
+    public async Task<ActionResult> CaregiverAccess(string profile, int code)
+    {
+      Family thisFamily = _db.Families.FirstOrDefault(family => family.ProfileName == profile);
+      // var isAuthorized = await _authService.AuthorizeAsync(User, thisFamily, YourVillageOperations.Access);
+      // if (!isAuthorized.Succeeded)
+      // {
+      //   return Forbid();
+      // }
+      var userId = this.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+      var currentUser = await _userManager.FindByIdAsync(userId);
+      if (thisFamily.SecretCode == code)
+      {
+        _db.CaregiverFamilies.Add(new CaregiverFamily() { CaregiverId = currentUser.Id, FamilyId = thisFamily.FamilyId });
+        // thisFamily.Caregivers.Add(new CaregiverFamily() { CaregiverId = currentUser.Id, FamilyId = thisFamily.FamilyId });
+        // thisFamily.CaregiverIds.Concat(currentUser.Id);
+        // _db.Families.Update(thisFamily);
+      }
+      else
+      {
+        return View();
+      }
+      _db.Entry(thisFamily).State = EntityState.Modified;
+      _db.SaveChanges();
+      return RedirectToAction("Index");
     }
 
     public ActionResult Edit(int id)
@@ -67,8 +160,13 @@ namespace YourVillage.Controllers
     }
 
     [HttpPost]
-    public ActionResult Edit(Family family)
+    public async Task<ActionResult> Edit(Family family)
     {
+      var isAuthorized = await _authService.AuthorizeAsync(User, family, YourVillageOperations.Update);
+      if (!isAuthorized.Succeeded)
+      {
+        return Forbid();
+      }
       _db.Entry(family).State = EntityState.Modified;
       _db.SaveChanges();
       return RedirectToAction("Index");
@@ -81,9 +179,14 @@ namespace YourVillage.Controllers
     }
 
     [HttpPost, ActionName("Delete")]
-    public ActionResult DeleteConfirmed(int id)
+    public async Task<ActionResult> DeleteConfirmed(int id)
     {
       var thisFamily = _db.Families.FirstOrDefault(family => family.FamilyId == id);
+      var isAuthorized = await _authService.AuthorizeAsync(User, thisFamily, YourVillageOperations.Delete);
+      if (!isAuthorized.Succeeded)
+      {
+        return Forbid();
+      }
       _db.Families.Remove(thisFamily);
       _db.SaveChanges();
       return RedirectToAction("Index");
